+++ tsp/src/instance.rs
@@ -30,13 +30,13 @@ mod tests {
 
     #[test]
     fn test_instance() {
-        let problem = PointsAndFunction::<f64, i64, Euc2d>::new(vec![
+        let problem = PointsAndFunction::<f64, f64, Euc2d>::new(vec![
             Point(0.0, 0.0),
             Point(0.0, 3.0),
             Point(4.0, 3.0),
         ]);
-        let solution = List::<PointsAndFunction<f64, i64, Euc2d>>::new((0..3).map(City).collect());
+        let solution = List::<PointsAndFunction<f64, f64, Euc2d>>::new((0..3).map(City).collect());
         let instance = Instance::new(problem, solution);
-        assert_eq!(instance.objective_value(), 12);
+        assert_eq!(instance.objective_value(), 12.0);
     }
 }
